pyproject.toml
[project]
name = "events-face"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.1.7",
    "django-filter>=25.1",
    "djangorestframework>=3.15.2",
    "dotenv>=0.9.9",
    "psycopg2-binary>=2.9.10",
    "ruff>=0.11.2",
]


[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
